feat(storage): add a common blobstore to store all blobs (#3906)
Currently, zot uses one of the existing repos as the master copy for a blob to
achieve dedupe, which complicates dedupe tracking logic. Furthermore, we
have a global storage lock which is becoming a bottleneck. In order to
move to a per-repo lock, we first need to simplify this logic.

Now use a single hidden global repo (_blobstore/) as a blob store instead.

+++ pkg/extensions/sync/destination.go
@@ -90,6 +90,12 @@ func (registry *DestinationRegistry) GetImageReference(repo, reference string) (
 func (registry *DestinationRegistry) CommitAll(repo string, imageReference ref.Ref) error {
 	tempImageStore := getImageStoreFromImageReference(repo, imageReference, registry.log)
 
+	if tempImageStore == nil {
+		registry.log.Error().Str("repo", repo).Msg("failed to get temp image store for sync commit")
+
+		return zerr.ErrLocalImgStoreNotFound
+	}
+
 	defer os.RemoveAll(tempImageStore.RootDir())
 
 	repoDir := path.Join(tempImageStore.RootDir(), repo)
